pyproject.toml
[build-system]
requires = ["setuptools>=72"]
build-backend = "setuptools.build_meta"

[project]
name = "sbnd"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.9",
    "lightning",
    "hydra-core>=1.3",
    "hydra-colorlog",
    "omegaconf",
    "numpy",
    "scipy",
    "h5py",
    "pandas",
    "tqdm",
    "tabulate",
    "wandb",
]

[project.scripts]
sbnd-train = "sbnd.train:main"
sbnd-test  = "sbnd.test:main"

[tool.setuptools]
package-dir = {"sbnd" = "src"}
packages    = ["sbnd"]

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
